move RenderFramerateLimiter
